Fix Api

+++ src/app/Modules/Language/FillLyric/Commands/Line/AddNextLineCmd.h
@@ -14,7 +14,7 @@ namespace FillLyric {
 
     private:
         LyricWrapView *m_view;
-        qlonglong m_index;
+        int m_index;
         CellList *m_newList;
     };
